requirements.txt
ultralytics
roboflow
paddlepaddle 
paddleocr
opencv-python
python-dotenv
pandas
streamlit